requirements.txt
# pip list output
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
cffi==1.17.1
distro==1.9.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
json_repair==0.39.1
neo4j==5.28.1
neo4j-graphrag==1.6.0
numpy==2.2.4
openai==1.68.0
pip==23.2.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pypdf==5.4.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
sniffio==1.3.1
sounddevice==0.5.1
tqdm==4.67.1
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2